feat: customise MessageSnackbar to replace BaseSnackbar
