--- src/BaseSnackbar.tsx
@@ -1,36 +1,0 @@
-import { ReactElement } from "react";
-import {
-  Snackbar,
-  SnackbarContent,
-  makeStyles,
-  ThemeProvider,
-} from "@material-ui/core";
-
-import { theme } from "./theme";
-
-const useStyle = makeStyles(() => ({
-  snackbar: {
-    background: theme.palette.info.light,
-  },
-}));
-
-interface Props {
-  open: boolean;
-  handleClose: () => void;
-  message: ReactElement;
-}
-
-function BaseSnackbar(props: Props): ReactElement {
-  const classes = useStyle();
-
-  return (
-    <ThemeProvider theme={theme}>
-      <Snackbar open={props.open} onClose={props.handleClose}>
-        <SnackbarContent className={classes.snackbar} message={props.message} />
-      </Snackbar>
-    </ThemeProvider>
-  );
-}
-
-export { BaseSnackbar };
-export type { Props as BaseSnackbarProps };
