+++ src/MessageSnackbar.tsx
@@ -1,7 +1,12 @@
 import { ReactElement } from "react";
 import SVG from "react-inlinesvg";
-import { IconButton, makeStyles } from "@material-ui/core";
-import { BaseSnackbar, BaseSnackbarProps } from "./BaseSnackbar";
+import {
+  IconButton,
+  makeStyles,
+  Snackbar,
+  SnackbarContent,
+  ThemeProvider,
+} from "@material-ui/core";
 import { theme } from "./theme";
 
 export const imgSrc = (src: string, type = "svg"): string =>
@@ -15,6 +20,7 @@ const useStyle = makeStyles(() => ({
     marginRight: "9px",
     marginTop: "0px",
     marginBottom: "-4px",
+    fill: theme.palette.primary.light,
   },
   message: {
     display: "inline-block",
@@ -34,41 +40,43 @@ const useStyle = makeStyles(() => ({
     marginBottom: "0px",
     fill: theme.palette.primary.light,
   },
+  snackbarContent: { backgroundColor: theme.palette.info.light },
 }));
 
-interface Props extends Omit<BaseSnackbarProps, "message"> {
+interface Props {
+  open: boolean;
+  onClose: () => void;
   messageText: string;
 }
 
-function MessageSnackbar({
-  open,
-  handleClose,
-  messageText,
-}: Props): ReactElement {
+function MessageSnackbar({ open, onClose, messageText }: Props): ReactElement {
   const classes = useStyle();
   return (
-    <BaseSnackbar
-      open={open}
-      handleClose={handleClose}
-      message={
-        <span>
-          <SVG src={imgSrc("warning")} className={classes.svgSmall} />
+    <ThemeProvider theme={theme}>
+      <Snackbar open={open} onClose={onClose}>
+        <SnackbarContent
+          className={classes.snackbarContent}
+          message={
+            <span>
+              <SVG src={imgSrc("warning")} className={classes.svgSmall} />
 
-          <div className={classes.message}>{messageText}</div>
-          <IconButton
-            size="small"
-            aria-label="close"
-            onClick={handleClose}
-            className={classes.iconButton}
-          >
-            <SVG
-              src={imgSrc("remove-label-icon")}
-              className={classes.svgSmallClose}
-            />
-          </IconButton>
-        </span>
-      }
-    />
+              <div className={classes.message}>{messageText}</div>
+              <IconButton
+                size="small"
+                aria-label="close"
+                onClick={onClose}
+                className={classes.iconButton}
+              >
+                <SVG
+                  src={imgSrc("remove-label-icon")}
+                  className={classes.svgSmallClose}
+                />
+              </IconButton>
+            </span>
+          }
+        />
+      </Snackbar>
+    </ThemeProvider>
   );
 }
